# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DisputeShield API
  version: 1.0.0
  description: "Embeddable dispute and SLA management for fintechs.\n\nThree surfaces, three authentication\
    \ schemes, deliberately not interchangeable:\n\n  * **Server API** — your backend, secret key. Mints\
    \ session tokens.\n  * **Widget API** — the browser, session token. Scoped to exactly one customer.\n\
    \  * **Management API** — your agents, an agent-scoped key plus an acting-agent header.\n\nTwo rules\
    \ apply everywhere and are worth stating once:\n\n  * **Cross-boundary access returns 404, never 403.**\
    \ A 403 confirms the resource\n    exists, which is an information leak (§8.1). This includes a role\
    \ that is not\n    permitted to use an endpoint — answering 403 there confirms the endpoint is real.\n\
    \  * **Every write requires `Idempotency-Key`.** Not optional: an optional key means\n    the guarantee\
    \ silently does not apply to the callers who most need it, and they\n    have no way to tell. Replaying\
    \ a key returns the original response; reusing one\n    with a different body is a 409, because returning\
    \ the first response would hide\n    a client bug.\n"
  license:
    name: BUSL-1.1
    url: https://github.com/nobledeveloper01/DisputeShield/blob/main/LICENSE
servers:
- url: https://api.disputeshield.dev
  description: Hosted
- url: http://localhost:8000
  description: Local
security:
- secretKey: []
components:
  securitySchemes:
    secretKey:
      type: http
      scheme: bearer
      description: '`ds_{env}_{random32}`. Server-side only. Argon2id-hashed at rest; shown

        once at creation and never retrievable. Scoped to one environment and one

        tenant — no key spans tenants.

        '
    sessionToken:
      type: http
      scheme: bearer
      description: '`dst_...`. Minted by your backend via POST /v1/sessions, scoped to exactly

        one customer_ref, 30-minute TTL, revocable (ADR-0002). The publishable key

        cannot substitute for this on any data endpoint.

        '
    publishableKey:
      type: http
      scheme: bearer
      description: '`pk_{env}_{random32}`. Safe to embed in a public page: it can read widget configuration
        and theming and nothing else. There is no code path from a publishable key to a dispute record,
        and the widget suite asserts that against every data route rather than a sample.'
    agentKey:
      type: http
      scheme: bearer
      description: 'A secret key, plus `X-DisputeShield-Acting-Agent` naming the person the call is made
        for. The audit trail records that person rather than the key: "an API key resolved this case"
        is not an answer a supervisor accepts.'
...
paths:
...
  /v1/widget/config:
    get:
      summary: Theme, categories and locale
      operationId: widgetConfig
      security:
      - publishableKey: []
      responses:
        '200':
          description: Configuration. Everything here is safe to disclose.
...